Subclass webview to override `canPerformAction`
This will stop those default `copy`, `cut` and `share` options from
coming up. Instead of using a default UIWebview we are subclassing it
and overriding `canPerformAction` to return NO. This will make this
webview to not add any options to the `UIMenuController` that pops up
when user selects a text. According to the responder chain, everything
will then be up the next view in heirarchy, in this case,
RCTWebViewBridge, where we already have the code to show our custom
actions -- `highlight` and `comment`, and nothing else.